requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
commonmark==0.9.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
executing==1.1.1
feedparser==6.0.10
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
traitlets==5.4.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.2.0
